pyproject.toml
[tool.poetry]
name = "python-project"
version = "0.1.0"
description = ""
authors = ["Caio Cardoso <imcaiofelipe@gmail.com>"]
readme = "README.md"
packages = [{include = "python_project"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.101.0"
uvicorn = {extras = ["standart"], version = "^0.23.2"}
pytest = "^7.4.0"
httpx = "^0.24.1"
sqlalchemy = "^2.0.20"
psycopg2 = "^2.9.7"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
